requirements.txt
cython
Flask
gunicorn
flask_restful
pyjnius
# For the clients
requests
